+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### Changed
+- Accept and ignore null as jsonldContext context param.
+
 ## [1.0.0] - 2015-04-08
 
 ### Changed